+++ src/Console/Daemon/DaemonCommand.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace CertaMesh\Gaze\Console\Daemon;
 
 use CertaMesh\Gaze\BinaryResolver;
+use CertaMesh\Gaze\Console\Concerns\BuildsBinaryArgv;
 use Illuminate\Console\Command;
 use Illuminate\Contracts\Config\Repository as ConfigRepository;
 
@@ -23,6 +24,8 @@ use Illuminate\Contracts\Config\Repository as ConfigRepository;
  */
 abstract class DaemonCommand extends Command
 {
+    use BuildsBinaryArgv;
+
     /**
      * Verb-specific flag list assembled from config + artisan options.
      *
@@ -49,12 +52,5 @@ abstract class DaemonCommand extends Command
         }
 
         return $argv;
-    }
-
-    protected function stringOption(string $name): ?string
-    {
-        $value = $this->option($name);
-
-        return is_string($value) && $value !== '' ? $value : null;
     }
 }